🚚 move other "domain" files to "segmentCollection"
