+++ packages/rum-recorder/src/domain/segmentCollection/deflateWorker.js
