+++ packages/rum-recorder/src/domain/segmentCollection/deflateWorker.spec.ts
