+++ packages/rum-recorder/src/domain/segmentCollection/segment.spec.ts
@@ -1,7 +1,7 @@
 import { noop, setDebugMode } from '@datadog/browser-core'
-import { isIE } from '../../../core/test/specHelper'
-import { MockWorker, parseSegment } from '../../test/utils'
-import { Record, RecordType, SegmentContext } from '../types'
+import { isIE } from '@datadog/browser-core/test/specHelper'
+import { MockWorker, parseSegment } from '../../../test/utils'
+import { Record, RecordType, SegmentContext } from '../../types'
 import { Segment } from './segment'
 
 const CONTEXT: SegmentContext = { application: { id: 'a' }, view: { id: 'b' }, session: { id: 'c' } }
