+++ packages/rum-recorder/src/domain/segmentCollection/segmentCollection.ts
@@ -1,7 +1,7 @@
 import { addErrorToMonitoringBatch, addEventListener, DOM_EVENT, EventEmitter, monitor } from '@datadog/browser-core'
 import { LifeCycle, LifeCycleEventType, ParentContexts, RumSession } from '@datadog/browser-rum-core'
-import { SEND_BEACON_BYTE_LENGTH_LIMIT } from '../transport/send'
-import { CreationReason, Record, SegmentContext, SegmentMeta } from '../types'
+import { SEND_BEACON_BYTE_LENGTH_LIMIT } from '../../transport/send'
+import { CreationReason, Record, SegmentContext, SegmentMeta } from '../../types'
 import { createDeflateWorker, DeflateWorker } from './deflateWorker'
 import { Segment } from './segment'
